refactor: consolidate token swap types and remove deprecated swap interface
