+++ packages/core-js/src/types/bff/index.ts
@@ -3,4 +3,3 @@ export * from './chains.js';
 export * from './common.js';
 export * from './tokens.js';
 export * from './estimate.js';
-export * from './swap.js';
